+++ tests/test_document.py
@@ -2999,7 +2999,7 @@ class DocumentTest(unittest.TestCase):
         self.assertEqual(User.objects.first(), bob)
         self.assertEqual(Book.objects.first(), hp)
 
-        # DeRefecence
+        # DeReference
         class AuthorBooks(Document):
             author = ReferenceField(User)
             book = ReferenceField(Book)
@@ -3026,6 +3026,18 @@ class DocumentTest(unittest.TestCase):
         self.assertEqual(User._get_collection(), get_db("testdb-1")[User._get_collection_name()])
         self.assertEqual(Book._get_collection(), get_db("testdb-2")[Book._get_collection_name()])
         self.assertEqual(AuthorBooks._get_collection(), get_db("testdb-3")[AuthorBooks._get_collection_name()])
+
+    def test_db_alias_propagates(self):
+        """db_alias propagates?
+        """
+        class A(Document):
+            name = StringField()
+            meta = {"db_alias": "testdb-1", "allow_inheritance": True}
+
+        class B(A):
+            pass
+
+        self.assertEquals('testdb-1', B._meta.get('db_alias'))
 
     def test_db_ref_usage(self):
         """ DB Ref usage in __raw__ queries """